fix proper return type for counter updates
the mzmetrics library returned int instead of the uint64_t which the
counter is based on. Moreover counter updates were also using int
instead of int64_t

+++ c_src/mzmetrics.c
@@ -10,7 +10,7 @@ void metrics_deinit(void)
     capi_deinit_counters();
 }
 
-counter64_t metrics_update_counter(struct GrpCounter *counter, unsigned int key, int val)
+counter64_t metrics_update_counter(struct GrpCounter *counter, unsigned int key, int64_t val)
 {
     return capi_update_counter(counter, key, val);
 }
